restructure the file tree and code of util functions
