+++ src/helpers/is-all-null.ts
@@ -1,0 +1,3 @@
+export function isAllNull(obj: Object) {
+  return Object.values(obj).every((value) => value === null);
+}
