+++ src/helpers/is-perceptor.ts
@@ -1,0 +1,2 @@
+export const isPerceptor = (): boolean =>
+  window.location.search.includes('?redirect=perceptor');
