+++ src/pages/ContentScripts/features/developer-networks/view.tsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react';
 import ReactModal from 'react-modal';
 
 import Graph from '../../../../components/Graph';
-import { getMessageByLocale } from '../../../../utils/utils';
+import { getMessageByLocale } from '../../../../helpers/get-message';
 import optionsStorage, {
   HypercrxOptions,
   defaults,
