+++ src/pages/ContentScripts/features/oss-gpt/index.tsx
@@ -3,7 +3,8 @@ import { render, Container } from 'react-dom';
 import $ from 'jquery';
 
 import features from '../../../../feature-manager';
-import { isPublicRepo, getGithubTheme } from '../../../../utils/utils';
+import { getGithubTheme } from '../../../../helpers/get-github-theme';
+import { isPublicRepo } from '../../../../helpers/is-public-repo';
 import { getRepoName } from '../../../../helpers/get-repo-info';
 import View from './view';
