+++ src/pages/ContentScripts/features/oss-gpt/view.tsx
@@ -11,7 +11,7 @@ import {
 import { getAnswer } from './service';
 import './rcw.scss';
 import exists from '../../../../helpers/exists';
-import { getMessageByLocale } from '../../../../utils/utils';
+import { getMessageByLocale } from '../../../../helpers/get-message';
 import optionsStorage, {
   HypercrxOptions,
   defaults,
