+++ src/pages/ContentScripts/features/perceptor-tab/index.tsx
@@ -3,7 +3,8 @@ import elementReady from 'element-ready';
 
 import iconSvgPath from './icon-svg-path';
 import features from '../../../../feature-manager';
-import { isPerceptor, isPublicRepo } from '../../../../utils/utils';
+import { isPerceptor } from '../../../../helpers/is-perceptor';
+import { isPublicRepo } from '../../../../helpers/is-public-repo';
 import sleep from '../../../../helpers/sleep';
 
 const featureId = features.getFeatureID(import.meta.url);
