+++ src/pages/ContentScripts/features/repo-issue-tooltip/view.tsx
@@ -1,16 +1,14 @@
 import React, { useState, useEffect } from 'react';
 
-import {
-  getGithubTheme,
-  getMessageByLocale,
-  isNull,
-  isAllNull,
-} from '../../../../utils/utils';
+import { getGithubTheme } from '../../../../helpers/get-github-theme';
+import { getMessageByLocale } from '../../../../helpers/get-message';
+import { isNull } from '../../../../helpers/is-null';
+import { isAllNull } from '../../../../helpers/is-all-null';
 import optionsStorage, {
   HypercrxOptions,
   defaults,
 } from '../../../../options-storage';
-import { generateDataByMonth } from '../../../../utils/data';
+import { generateDataByMonth } from '../../../../helpers/data';
 import ReactTooltip from 'react-tooltip';
 import IssueChart from './IssueChart';
