+++ src/pages/ContentScripts/features/repo-star-tooltip/view.tsx
@@ -1,11 +1,12 @@
 import React, { useState, useEffect } from 'react';
 
-import { getGithubTheme, getMessageByLocale } from '../../../../utils/utils';
+import { getMessageByLocale } from '../../../../helpers/get-message';
+import { getGithubTheme } from '../../../../helpers/get-github-theme';
 import optionsStorage, {
   HypercrxOptions,
   defaults,
 } from '../../../../options-storage';
-import { generateDataByMonth } from '../../../../utils/data';
+import { generateDataByMonth } from '../../../../helpers/data';
 import ReactTooltip from 'react-tooltip';
 import StarChart from './StarChart';
